==> 01_save_uniqueKeys.ipynb <==
import pickle
import pandas as pd
import ast
from collections import OrderedDict
import unicodedata
from ftfy import fix_text
import re
import string

ETSY_ROUND = '_2019'
SOURCE_FILE = ETSY_ROUND + '/source/producs_2017_2019.pickle'
OUTPUT_FOLDER = ETSY_ROUND + '/output/'

df_source = pd.read_pickle(SOURCE_FILE)

#df_source.head()

df = df_source[['listing_id', 'materials']]

df.tail(2)

def removeAnyRepeatedSpace(texte):
    return re.sub(' +',' ', texte.strip())

# removeAnyRepeatedSpace(' stri   ng ')
# 'stri ng'

def removePunctuation(texte):
    exclude_punt = set(string.punctuation)
    texte = ''.join(ch for ch in texte if ch not in exclude_punt)
    return str(texte)

def removeAccents(texte):
    texte = unicodedata.normalize('NFD', texte)
    texte = texte.encode('ascii', 'ignore')
    texte = texte.decode("utf-8")
    return str(texte)

data_dict = OrderedDict()
for row in df_source.itertuples():
    index = 1
    try:
        materials = ast.literal_eval(row.materials)
        for material in materials:
            idx = '{}_{}'.format(row.listing_id, index)

            material_clean = removePunctuation(removeAnyRepeatedSpace(fix_text(material))).strip().lower()

            material_key = removeAccents(material_clean).replace(' ', '_')

            data_dict[idx] = [row.listing_id, material_key, material_clean]
            index = index + 1
    except:
        print('Error: {}'.format(row.materials))

# data_dict

pickle.dump( data_dict, open( OUTPUT_FOLDER + "data_dict.pickle", "wb" ) )

dic_uniques = OrderedDict()

for (key, value) in data_dict.items() :
    if value[2] not in dic_uniques:
        dic_uniques[value[1]] = value[2]

len(dic_uniques)

# dic_uniques

pickle.dump( dic_uniques, open( OUTPUT_FOLDER + "dic_uniques.pickle", "wb" ) )
