==> 02_saveKeysToTransalte.ipynb <==
import re
import enchant
import pickle
import pandas as pd

ETSY_ROUND = '_2019'
OUTPUT_FOLDER = ETSY_ROUND + '/output/'

dictionary_FR = enchant.Dict("fr_FR")
dictionary_GB = enchant.Dict("en_GB")
dictionary_US = enchant.Dict("en_US")

def isFrenchWord(word):
    return dictionary_FR.check(word)

def isEnglighWord(word):
    return dictionary_GB.check(word) or dictionary_US.check(word)

# isFrenchWord('maison')

def shouldWeTranslate(texte):
    words = texte.split()
    for word in words:
        if word.isalpha() and isFrenchWord(word) and not isEnglighWord(word):
            return True
    return False

# shouldWeTranslate("house 12px  ")
# False
# shouldWeTranslate("maison 12px  ")
# True

dic_uniques = pickle.load( open( OUTPUT_FOLDER + "dic_uniques.pickle", "rb" ) )
dic_to_translate = dict()

idx = 0
for (key, texte) in dic_uniques.items() :
    if shouldWeTranslate(texte):
        dic_to_translate[key] = texte

len(dic_to_translate)

dic_to_translate

pickle.dump( dic_to_translate, open( OUTPUT_FOLDER + "dic_to_translate.pickle", "wb" ) )

df = pd.DataFrame.from_dict(dic_to_translate, orient='index')

df.to_excel(OUTPUT_FOLDER + 'uniques_to_translate.xlsx')
