==> 03_translate.ipynb <==
import pickle
import pandas as pd
import time
from py_translator import Translator
translator = Translator()

ETSY_ROUND = '_2019'
OUTPUT_FOLDER = ETSY_ROUND + '/output/'

dic_uniques = pickle.load( open( OUTPUT_FOLDER + "dic_to_translate.pickle", "rb" ) )

df = pd.DataFrame.from_dict(dic_uniques, orient='index')

df.columns = ['materials']

df.info()

df.head()

dic_list = df['materials'].values.tolist()

dic_list[:10]

len(dic_list)

def chunks(l, n):
    """Yield successive n-sized chunks from l."""
    for i in range(0, len(l), n):
        yield l[i:i + n]

chunk_value = 120
google_max = 3900

gen = chunks(dic_list, chunk_value)

source_texts = []

for idx in range(1, int((len(dic_list)/chunk_value) + 1)):
    txt = '\n'.join(next(gen))
    if (len(txt)) >= google_max:
        print(len(txt))
    source_texts.append(txt)

len(source_texts)

def translateList(liste):
    translated_texts = []
    for index, txt in enumerate(liste):
        txt_translated = translator.translate(txt)
        if txt_translated == 'TRANSLATION_ERROR':
            print('TRANSLATION_ERROR {}'.format(index))
        else:
            print('OK {}'.format(index))
        translated_texts.append(txt_translated)
    return translated_texts

translated_texts = translateList(source_texts[0:50])

pickle.dump( translated_texts, open( OUTPUT_FOLDER + "translated_texts_0_50.pickle", "wb" ) )

translated_texts = translateList(source_texts[50:100])

pickle.dump( translated_texts, open( OUTPUT_FOLDER + "translated_texts_50_100.pickle", "wb" ) )

translated_texts = translateList(source_texts[100:150])

pickle.dump( translated_texts, open( OUTPUT_FOLDER + "translated_texts_100_150.pickle", "wb" ) )

translated_texts = translateList(source_texts[150:])

pickle.dump( translated_texts, open( OUTPUT_FOLDER + "translated_texts_150_207.pickle", "wb" ) )
