==> 04_mergeTranslated.ipynb <==
import pickle

ETSY_ROUND = '_2019'
OUTPUT_FOLDER = ETSY_ROUND + '/output/'

translated_0_50 = pickle.load( open( OUTPUT_FOLDER + "translated_texts_0_50.pickle", "rb" ) )
translated_50_100 = pickle.load( open( OUTPUT_FOLDER + "translated_texts_50_100.pickle", "rb" ) )
translated_100_150 = pickle.load( open( OUTPUT_FOLDER + "translated_texts_100_150.pickle", "rb" ) )
translated_150_207 = pickle.load( open( OUTPUT_FOLDER + "translated_texts_150_207.pickle", "rb" ) )

translated_list = [*translated_0_50, *translated_50_100, *translated_100_150, *translated_150_207]

len(translated_list)

translated_list_text = [t.text for t in translated_list]

translated_list_text[:50]

pickle.dump( translated_list_text, open( OUTPUT_FOLDER + "translated_list.pickle", "wb" ) )
