==> 05_mergeOrigin.ipynb <==
import pickle
from collections import OrderedDict
import pandas as pd

ETSY_ROUND = '_2019'
OUTPUT_FOLDER = ETSY_ROUND + '/output/'

translated_list = pickle.load( open( OUTPUT_FOLDER + "translated_list.pickle", "rb" ) )

list_uniques = []

for value in translated_list:
    texts = value.split('\n')
    for text in texts:
        list_uniques.append(text.strip())

list_uniques

len(list_uniques)

df = pd.DataFrame.from_dict(list_uniques)

df.head()

df.to_excel(OUTPUT_FOLDER + 'uniques_translated.xlsx')
