==> 06_to_final_dic.ipynb <==
import pandas as pd

ETSY_ROUND = '_2019'
OUTPUT_FOLDER = ETSY_ROUND + '/output/'

df_fr = pd.read_excel(OUTPUT_FOLDER + 'uniques_to_translate.xlsx')
df_en = pd.read_excel(OUTPUT_FOLDER + 'uniques_translated.xlsx')

df_fr.head()

df_en.head()

df_fr.columns = ['key', 'fr']
df_fr.head()

df_fr['en'] = df_en[0]

df_fr.head()

df_final = df_fr.set_index(['key'])

df_final.head()

df_final.to_excel(OUTPUT_FOLDER + 'uniques_translate_final.xlsx')
