==> 07_to_origin.ipynb <==
import pickle
import pandas as pd
from collections import OrderedDict
import spacy
nlp = spacy.load('en')

ETSY_ROUND = '_2019'
OUTPUT_FOLDER = ETSY_ROUND + '/output/'

df = pd.read_excel(OUTPUT_FOLDER + 'uniques_translate_final.xlsx')

noisy_pos_tags = ["PROP","DET","PART","CCONJ","ADP","PRON","VERB","ADJ"]
min_token_length = 2 # minimum token length to remove

def is_noise(token):
    '''
    standard way to validate spacy tokens
    This method validate all the passed tokens and set true false on it
    '''
    is_noise = False
    if token.pos_ in noisy_pos_tags:
        is_noise = True
    elif token.is_stop == True:
        is_noise = True
    elif token.is_digit == True:
        is_noise = True
    elif token.is_space == True:
        is_noise = True
    elif len(token.string) <= min_token_length:
        is_noise = True
    return is_noise

def lemmanize(string):
    if isinstance(string, str):
        result_tempo = []
        nlp_words = nlp(string)
        for word in nlp_words:
            if not is_noise(word):
                result_tempo.append(word.lemma_)
        if len(result_tempo) > 0:
            return ' '.join(result_tempo)
        else:
            return ''
    else:
        print(string)

dic_translated = dict()

for index, row in df.iterrows():
    dic_translated[row.key] = row.en

dic_translated

dic_uniques = pickle.load( open( OUTPUT_FOLDER + "dic_uniques.pickle", "rb" ) )

for (key, texte) in dic_uniques.items() :
    if key in dic_translated:
        dic_uniques[key] = dic_translated[key]

dic_uniques['corail_rouge']

len(dic_uniques)

dic_uniques_lem = OrderedDict()

for (key, texte) in dic_uniques.items() :
        dic_uniques_lem[key] = [dic_uniques[key], lemmanize(dic_uniques[key])]

dic_uniques_lem['corail_rouge']

pickle.dump( dic_uniques_lem, open( OUTPUT_FOLDER + "dic_uniques_lem.pickle", "wb" ) )

data_dict = pickle.load( open( OUTPUT_FOLDER + "data_dict.pickle", "rb" ) )

len(data_dict)

data_dict

dict_final = OrderedDict()

for (key, row) in data_dict.items() :
    id_listing = row[0]
    key_texte = row[1]
    translated_texte = dic_uniques_lem[key_texte][0]
    lemmanized_texte = dic_uniques_lem[key_texte][1]
    dict_final[key] = [id_listing, key_texte, translated_texte, lemmanized_texte]

data_final = []

idx = 0
for (key, row) in dict_final.items():
    data_final.append(row)
    idx = idx + 1

df = pd.DataFrame(data_final, columns=['id_listing', 'key_texte', 'translated_texte', 'lemmanized_texte'])

df.head()

df.to_csv(OUTPUT_FOLDER + 'df_final.csv')
