==> 08_stats.ipynb <==
import pandas as pd

ETSY_ROUND = '_2019'
OUTPUT_FOLDER = ETSY_ROUND + '/output/'

df = pd.read_csv(OUTPUT_FOLDER + 'df_final.csv')

len(df)

df.head()

df = df.dropna()

len(df)

df = df.drop(['Unnamed: 0'], axis = 1).set_index(['id_listing'])

df.head()

df.to_csv(OUTPUT_FOLDER + 'df_final_no_na.csv')

df_counts = df.groupby(['lemmanized_texte']).size().reset_index()

df_counts.head()

df_counts.rename(columns={0 : 'counts'}, inplace=True)

df_counts.head()

df_counts = df_counts.sort_values('counts',ascending=False)

df_counts.head(5)

len(df)
