==> 09_toSqlite.ipynb <==
import pandas as pd
from sqlalchemy import create_engine

ETSY_ROUND = '_2019'
OUTPUT_FOLDER = ETSY_ROUND + '/output/'
SOURCE_FILE = ETSY_ROUND + '/source/producs_2017_2019.pickle'

df_products = pd.read_pickle(SOURCE_FILE)

df_materials = pd.read_csv(OUTPUT_FOLDER + 'df_final_no_na.csv')

# Create database
disk_engine = create_engine('sqlite:///{}ETSY_3.db'.format(OUTPUT_FOLDER))

df_products.to_sql('products', disk_engine, if_exists='replace')

df_materials.to_sql('materials', disk_engine, if_exists='replace')
